+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All versions prior to 1.0.0 are untracked.
 - Added tests for verifying signatures created with v0.3.1
 - cli: `model_signing sign` now supports the `--oauth_force_oob` option (default: False)
 - Added support for specifying `--client_id` and `--client_secret` for OIDC authentication.
+- cli: Added support for `--allow_symlinks` option
 
 ## [1.0.1] - 2024-04-18